Module Tezt_cloud.Alert

type receiver

val slack_receiver : ?channel:string -> name:string -> api_url:string -> unit -> receiver

val null_receiver : receiver

type route

val route : ?group_wait:string -> ?group_interval:string -> ?repeat_interval:string -> receiver -> route

route ?group_wait ?group_interval ?repeat_interval receiver creates a fresh route whose receiver is receiver and where:

group_wait defines how long to wait before sending a notification about new alerts. If omitted, inherit the group_wait of the default route.

group_interval defines how long to wait before sending notification about new alerts for a group. If omitted, inherit the group_interval of the default route.

repeat_interval defines the minimum time interval between sending two notifications about the same alert. If omitted, inherit the repeat_interval of the default route. The repeat_interval value should be a multiple of group_interval.

type severity =

| Critical

| Warning

| Info

Severity of an alert.

type t

Type of an alert.

val make : ?route:route -> ?for_:string -> ?description:string -> ?summary:string -> ?severity:severity -> ?group_name:string -> ?interval:string -> name:string -> expr:string -> unit -> t

make ?route ?for_ ?description ?summary ?severity ~name ~expr defines a new Prometheus alert with name name and promQL expr. Optionally a severity, summary and description of the alert can be defined.

If route is provided, the alert can be routed to a receiver (Slack, webhook, ...).