Module Tezt_cloud.Alert_manager

type alert

Prometheus alert.

val alert : expr:string -> severity:[ `Critical | `Info | `None | `Warning ] -> ?for_:string -> ?description:string -> ?summary:string -> string -> alert

alert name ~expr ~severity ?for_ ?description ?summary creates an alert with:

name: alert identifier.

expr: promQL expression that triggers the alert.

severity: alert level (either `Critical, `Warning, `Info or `None).

for_: optional duration before firing.

description: optional detailed explanation.

summary: optional brief overview.

val name_of_alert : alert -> string

name_of_alert alert returns the name associated with the alert.

val alert_template : alert -> Jingoo.Jg_types.tvalue

Jingoo alert template.

type config

Configuration type for Alert manager.

val add_global : k:string -> v:string -> config -> config

add_global k v config binds key k to value v in the configuration globals of config. Globals is a map of key-value pairs that defines global settings like webhook URLs and authentication tokens that apply across all alerts and receivers.

val add_globals : (string * string) list -> config -> config

add_globals kvs config is add_global for a list of key-value pairs.

type receiver

A receiver defines how to send notifications (like to Slack, email, etc.) when an alert is triggered, including the destination and message format.

val add_receiver : receiver -> config -> config

add_receiver receiver config extends the config with the provided receiver.

val slack_receiver : name:string -> channel:string -> title:string -> text:string -> receiver

slack_receiver ~name ~channel ~text config creates a receiver named name, sending notifications on the slack channel, with the given title and text metadata.

val slack_webhook_receiver : name:string -> receiver

slack_webhook_receiver ~name creates a receiver named name, sending notifications on the channel configured on the slack side. This is a hack implemented in order to use slack receivers with modern slack webhooks in alert manager. More precisely the channel field of the payload received by slack will be the description of the alert such that it can be bound in the slack message sent by the webhook.

type group

A group defines when and how alerts are processed, including timing parameters (group_wait, group_interval, repeat_interval), which alerts to match (alert_names), and where to send them (receiver_name).

val group : ?group_wait:string -> ?group_interval:string -> ?repeat_interval:string -> string -> receiver -> group

group name receiver ?group_wait ?group_interval ?repeat_interval creates a fresh group named named whose receiver is receiver and where:

group_wait defines how long to wait before sending a notification about new alerts. If omitted, child routes inherit the group_wait of the parent route.

group_interval defines how long to wait before sending notification about new alerts for a group. If omitted, child routes inherit the group_interval of the parent route.

repeat_interval defines the minimum time interval between sending two notifications about the same alert. If omitted, child routes inherit the repeat_interval of the parent route. The repeat_interval value should be a multiple of group_interval.

val add_group : group -> config -> config

add_group group config extends the configuration config with the provided group.

val with_group : group -> (group -> group) -> config -> config

with_group group f config updates the group in config by applying f. The group must be part of the config.

val add_group_alert : alert -> group -> group

add_group_alert alert group updates the given group to match against alert.

val add_group_alerts : alert list -> group -> group

add_group_alerts alerts group is add_group_alert for a list of alerts.

module Handler : sig ... end

Alert handler.

module Collection : sig ... end

A collection is a registery of alerts and handlers.

type t

The alert manager service.

val run : Collection.t -> string list -> t Lwt.t

run collection alert_handlers starts a new docker container in the background, using the latest Prometheus Alert Manager image. Activates the given alert_handlers. Note that each alert handler name must be registered in collection.

val shutdown : unit -> unit Lwt.t

shutdown t shutdowns the alert manager service.