Module Alert_manager.Collection

A collection is a registery of alerts and handlers.

type t

The collection type.

val init : unit -> t

init () returns a fresh, empty collection.

val register_alert : alert -> t -> unit

register_alert alert t registers the given alert in the collection t.

val register_handler : Handler.t -> t -> unit

register_handler handler t registers the given handler in the collection t.

val alerts : t -> alert list

alerts t returns the alerts registered in collection t.

val handlers : t -> Handler.t list

handlers t returns the handlers registered in collection t.