Module Alert_manager.Handler

Alert handler.

type t = {

name : string;

setup : config -> config;

}

Alert handler type where:

name is the handler name.

setup defines how the alter manager configuration should be updated when this handler is activated.

val make : name:string -> setup:(config -> config) -> t

make_handler ~name ~setup returns an handler whose name is name and setup function is setup.