Module Tezt_cloud.Cloud

type t

val register : ?proxy_files:string list -> ?proxy_args:string list -> ?vms:Configuration.t list -> __FILE__:string -> title:string -> tags:string list -> ?seed:Tezt.Test.seed -> ?alerts:Alert.t list -> (t -> unit Lwt.t) -> unit

A wrapper around Test.register that can be used to register new tests using VMs provided as a map indexed by name. Each VM is abstracted via the Agent module.

proxy_files should contains file that are needed by the scenario to run (only used for proxy mode).

proxy_args should contains CLI arguments necessary for the proxy mode. This can be used for example when an argument is provided via an environment variable instead of a command-line argument.

val push_metric : t -> ?help:string -> ?typ:[ `Counter | `Gauge ] -> ?labels:(string * string) list -> name:string -> float -> unit

push_metric t ?help ?typ ?labels ~name v pushes the value v for metric on Prometheus. labels can be used to categorise the metric (each set of label define a single curve). typ can be used to provide the type of the metric. help can be used to provide some naive documentation about the metrics.

val agents : t -> Agent.t list

agents t returns the list of agents deployed.

type target = {

agent : Agent.t;

port : int;

app_name : string;

}

val add_prometheus_source : t -> ?metrics_path:string -> name:string -> target list -> unit Lwt.t

add_prometheus_source t ?metrics_path ~name targets allows to add a new source of metrics that Prometheus can scrap. By default metric_path is /metrics. job_name is just the name to give for the job that will scrap the metrics. It must be unique. A target enables to define a list of points to scrap. Each point can have a name defined by app_name.

val add_service : t -> name:string -> url:string -> unit Lwt.t

val open_telemetry_endpoint : t -> string option